image_manager/requirements.txt
pillow>=10.0.0
tkinter-page>=0.0.5
watchdog>=3.0.0
sqlalchemy>=2.0.0